Tidying up rose_op2parloop:
 * Removal of pointless default constructors
 * Swapping tabs for spaces
 * Removal of commented code, and a function that was a no-op
 * Removal of incorrect/pointless comments

In removing all the comments that state "by Zohirul", I am not attempting to erase the evidence of his contribution to this project. There is no need for notes that state who wrote what in the source code, when this information is tracked by a revision control system.

To be clear, the majority of this code is written by Zohirul, based upon the minimal prototype that I produced on Jan 26 2010.